fix(hyprland): CopyQ is not reliable in Hyprland
Most of the time, I need to press the copy key multiple times before it
registers in CopyQ, and sometimes I need to move the cursor to the top
in the CopyQ 'show history' to make it work.

The combination of 'cliphist' and 'wl-clipboard' works well here. It is
very reliable in Hyprland. However, I can't find a runner that is able to
show a clipboard history with an image preview feature.

I use audio feedback for both of them. That's why I know when my text has
not landed in the primary clipboard.